% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_portfolio_flags.R
\name{add_portfolio_flags}
\alias{add_portfolio_flags}
\title{Adds flags columns to a portfolio data frame}
\usage{
add_portfolio_flags(portfolio, currencies)
}
\arguments{
\item{portfolio}{A data frame containing a portfolio}

\item{currencies}{A data frame containing a currencies and their exchange
rates.}
}
\value{
A data frame of the portfolio with the flags columns added
}
\description{
Adds flags columns to a portfolio data frame signifying if each holding has a
currency specified, has a valid value, has a valid ISIN, and has matching
financial data.
}
