% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_revenue_split.R
\name{add_revenue_split}
\alias{add_revenue_split}
\title{A short description of the function}
\usage{
add_revenue_split(has_revenue, portfolio, revenue_data)
}
\arguments{
\item{has_revenue}{A description of the argument}

\item{portfolio}{A description of the argument}

\item{revenue_data}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
