% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_portfolio_financed_emissions.R
\name{calculate_portfolio_financed_emissions}
\alias{calculate_portfolio_financed_emissions}
\title{Calculate the financed emissions of a portfolio, by sector and asset type}
\usage{
calculate_portfolio_financed_emissions(
  data,
  entity_info,
  entity_emission_intensities,
  average_sector_emission_intensities
)
}
\arguments{
\item{data}{A dataframe, like the total_portfolio object in
web_tool_script_1.}

\item{entity_info}{A dataframe, like the \code{entity_info} output of
\code{pacta.data.preparation}.}

\item{entity_emission_intensities}{A dataframe, like the
\code{entity_emission_intensities} output of \code{pacta.data.preparation}.}

\item{average_sector_emission_intensities}{A dataframe, like the
\code{average_sector_emission_intensities} output of \code{pacta.data.preparation}.}
}
\value{
A summarized dataset, indicating the portfolios financed emissions,
per sector and asset type.
}
\description{
This functions takes as input a total_portfolio object, and calculates the
financed emissions associated to that portfolio, by sector and asset type. If
no real emissions data exist for an entity, a sectoral average is used.
}
