% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_audit_information_data.R
\name{export_audit_information_data}
\alias{export_audit_information_data}
\title{Export audit data for Transition Monitor website}
\usage{
export_audit_information_data(
  audit_file_ = .GlobalEnv$audit_file,
  portfolio_total_ = .GlobalEnv$portfolio_total,
  folder_path = .GlobalEnv$proc_input_path,
  project_name_ = NA
)
}
\arguments{
\item{audit_file_}{A data frame containing the audit data}

\item{portfolio_total_}{A data frame containing the processed portfolio data}

\item{folder_path}{A length one character vector specifying the path where
the exported audit files should be saved}

\item{project_name_}{An optional length one character vector specifying a
project name to prepend to the filenames. Default is \code{NA} which has no
effect.}
}
\value{
Returns \code{NULL} invisibly as it is called only for its side effect of
saving files
}
\description{
Exports three audit files for the Transition Monitor website, allowing the
website GUI to report to the user details about the audit process result.
}
