% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_files.R
\name{get_input_files}
\alias{get_input_files}
\title{A short description of the function}
\usage{
get_input_files(portfolio_name_ref_all)
}
\arguments{
\item{portfolio_name_ref_all}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
