% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.portfolio.audit-package.R
\docType{package}
\name{pacta.portfolio.audit-package}
\alias{pacta.portfolio.audit}
\alias{pacta.portfolio.audit-package}
\title{pacta.portfolio.audit: Functions for Calculating Audit Statistics for PACTA on Capital Transition Monitor}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

These utils include functions that facilitate calculation of useful audit statistics for PACTA on Capital Transition Monitor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMI-PACTA/pacta.portfolio.audit}
  \item Report bugs at \url{https://github.com/RMI-PACTA/pacta.portfolio.audit/issues}
}

}
\author{
\strong{Maintainer}: CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item Rocky Mountain Institute \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
