% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_summary.R
\name{portfolio_summary}
\alias{portfolio_summary}
\title{A short description of the function}
\usage{
portfolio_summary(
  portfolio_total,
  grouping_variables = .GlobalEnv$grouping_variables
)
}
\arguments{
\item{portfolio_total}{A description of the argument}

\item{grouping_variables}{A character vector specifiying the grouping variables.
Typically \code{c("investor_name", "portfolio_name")}.}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
