% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_portfolio.R
\name{process_raw_portfolio}
\alias{process_raw_portfolio}
\title{A short description of the function}
\usage{
process_raw_portfolio(
  portfolio_raw,
  fin_data,
  fund_data,
  entity_info,
  currencies,
  grouping_variables,
  total_fund_list = NULL,
  isin_to_fund_table = isin_to_fund_table
)
}
\arguments{
\item{portfolio_raw}{A description of the argument}

\item{fin_data}{A description of the argument}

\item{fund_data}{A description of the argument}

\item{entity_info}{A description of the argument}

\item{currencies}{A description of the argument}

\item{grouping_variables}{A description of the argument}

\item{total_fund_list}{A description of the argument}

\item{isin_to_fund_table}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
