% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ald_flag.R
\name{create_ald_flag}
\alias{create_ald_flag}
\title{A short description of the function}
\usage{
create_ald_flag(portfolio, comp_fin_data, debt_fin_data)
}
\arguments{
\item{portfolio}{A description of the argument}

\item{comp_fin_data}{A description of the argument}

\item{debt_fin_data}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
