% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audit_file.R
\name{create_audit_file}
\alias{create_audit_file}
\title{Create the audit file for a processed portfolio}
\usage{
create_audit_file(portfolio_total, has_revenue)
}
\arguments{
\item{portfolio_total}{A data frame}

\item{has_revenue}{Logical determining if revenue data is included or not}
}
\value{
A data frame
}
\description{
Creates an audit file from a processed portfolio, selecting the expected
columns.
}
