% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audit_file.R
\name{create_audit_file}
\alias{create_audit_file}
\title{A short description of the function}
\usage{
create_audit_file(portfolio_total, grouping_variables, has_revenue)
}
\arguments{
\item{portfolio_total}{A description of the argument}

\item{grouping_variables}{A character vector defining grouping variable}

\item{has_revenue}{Logical determining if revenue data is included or not}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
