% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_audit_information_data.R
\name{export_audit_information_data}
\alias{export_audit_information_data}
\title{A short description of the function}
\usage{
export_audit_information_data(
  audit_file_ = .GlobalEnv$audit_file,
  portfolio_total_ = .GlobalEnv$portfolio_total,
  folder_path = .GlobalEnv$proc_input_path,
  project_name_ = NA
)
}
\arguments{
\item{audit_file_}{A description of the argument}

\item{portfolio_total_}{A description of the argument}

\item{folder_path}{A description of the argument}

\item{project_name_}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
