% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_info.R
\name{get_entity_info}
\alias{get_entity_info}
\title{Loads entity info so that it's ready to use}
\usage{
get_entity_info(dir = .GlobalEnv$analysis_inputs_path)
}
\arguments{
\item{dir}{A string containing the path to the directory where the entity
info files should be found, typically the value of the
\code{analysis_inputs_path} parameter}
}
\value{
A data frame containing the entity info, or a path with the class "db_path", either of which can be used as input to \code{left_join_entity_info()}
}
\description{
Loads entity info from the specified path so that it's ready to use, either
by loading the data into memory from an RDS file, or returning a path to an
SQLite file with the class "db_path".
}
