% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_info.R
\name{get_entity_info}
\alias{get_entity_info}
\title{A short description of the function}
\usage{
get_entity_info()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
