% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_files.R
\name{get_input_files}
\alias{get_input_files}
\title{Import a portfolio CSV according to parameters}
\usage{
get_input_files(
  portfolio_name_ref_all,
  project_location = .GlobalEnv$project_location
)
}
\arguments{
\item{portfolio_name_ref_all}{A length one character vector specifying the
portfolio name}

\item{project_location}{A character string defining the path to the user's
data directory, usually "working_dir" and the value held in the
\code{project_location} parameter}
}
\value{
A data frame containing the imported portfolio data
}
\description{
Import a portfolio CSV from a specified location after verifying that it
matches the specified \code{portfolio_name_ref_all} and a matching parameters file
exists.
}
