% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_files.R
\name{get_input_files}
\alias{get_input_files}
\title{A short description of the function}
\usage{
get_input_files(
  portfolio_name_ref_all,
  project_location = .GlobalEnv$project_location
)
}
\arguments{
\item{portfolio_name_ref_all}{A description of the argument}

\item{project_location}{A character string defining the path to the user's
data directory, usually "working_dir" and the value held in the
\code{project_location} parameter}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
