% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_summary.R
\name{portfolio_summary}
\alias{portfolio_summary}
\title{A short description of the function}
\usage{
portfolio_summary(portfolio_total, ...)
}
\arguments{
\item{portfolio_total}{A description of the argument}

\item{...}{A catch-all for deprecated variables (e.g. \code{grouping_variables})}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
